# contracts/http/openapi.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "EcoBin P0 HTTP API",
    "version": "1.0.0-rc.1",
    "description": "EcoBin /api/v1 transport contract. Business slices add their own paths and schemas without redefining these cross-cutting rules."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment"
    }
  ],
...
  "paths": {
    "/api/v1/web/auth/csrf-token": {
      "get": {
        "operationId": "getWebCsrfToken",
        "tags": [
          "Web authentication"
        ],
        "summary": "Bootstrap or rotate the SPA CSRF token",
        "security": [],
        "responses": {
          "200": {
            "description": "CSRF token returned; the response is never cacheable",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfTokenEnvelope"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalProblem"
          }
        }
      }
    },
    "/api/v1/web/auth/sessions": {
      "post": {
        "operationId": "createWebSession",
        "tags": [
          "Web authentication"
        ],
        "summary": "Create a tenant or staff Web session",
        "security": [
          {
            "webCsrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebLoginRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/WebSessionCreated"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/web/platform/auth/sessions": {
      "post": {
        "operationId": "createPlatformWebSession",
        "tags": [
          "Web authentication"
        ],
        "summary": "Create a platform administrator Web session",
        "security": [
          {
            "webCsrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebLoginRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/WebSessionCreated"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/web/auth/sessions/current": {
      "get": {
        "operationId": "getCurrentWebSession",
        "tags": [
          "Web authentication"
        ],
        "summary": "Read the current tenant or staff Web session",
        "security": [
          {
            "webSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/WebSessionCurrent"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentWebSession",
        "tags": [
          "Web authentication"
        ],
        "summary": "Revoke the current tenant or staff Web session",
        "security": [
          {
            "webSessionCookie": [],
            "webCsrfToken": []
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/CsrfProblem"
          }
        }
      }
    },
    "/api/v1/web/platform/auth/sessions/current": {
      "get": {
        "operationId": "getCurrentPlatformWebSession",
        "tags": [
          "Web authentication"
        ],
        "summary": "Read the current platform Web session",
        "security": [
          {
            "webSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/WebSessionCurrent"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentPlatformWebSession",
        "tags": [
          "Web authentication"
        ],
        "summary": "Revoke the current platform Web session",
        "security": [
          {
            "webSessionCookie": [],
            "webCsrfToken": []
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/CsrfProblem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "webSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-ecobin-web-session",
        "description": "Secure, HttpOnly, SameSite=Lax, Path=/ and no Domain. JavaScript never reads this value."
      },
      "webCsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN",
        "description": "Bootstrapped from GET /api/v1/web/auth/csrf-token and rotated after login/logout."
      },
      "miniappBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Exactly one short-lived aud=miniapp or aud=miniapp-staff session. No refresh token."
      },
      "wechatPaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Signature"
      },
      "wechatPayTimestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Timestamp"
      },
      "wechatPayNonce": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Nonce"
      },
      "wechatPaySerial": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Serial"
      }
    },
...
    "headers": {
      "NoStore": {
        "description": "Sensitive or mutable projections are not cacheable",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      },
      "RequestId": {
        "description": "Transport request identifier; never a business idempotency identity",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
...
      "RetryAfter": {
        "description": "Whole seconds before the client should retry",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      }
    },
    "responses": {
...
      "WebSessionCreated": {
        "description": "A new server-side Web session was created and the HttpOnly cookie was set",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WebSessionEnvelope"
            }
          }
        }
      },
      "WebSessionCurrent": {
        "description": "Current Web session and live capability projection",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WebSessionEnvelope"
            }
          }
        }
      },
...
      "NoContent": {
        "description": "The synchronous local transition completed and there is no response body",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
      "InvalidRequest": {
        "description": "Malformed JSON, type, format, or cross-field input",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "UnauthorizedProblem": {
        "description": "Authentication is missing, invalid, revoked, expired, or has the wrong audience",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "ForbiddenProblem": {
        "description": "The visible resource exists but the authenticated subject lacks the current capability",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
...
      "CsrfProblem": {
        "description": "The Web CSRF token is missing, stale, or invalid",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "example": {
              "code": "SECURITY.CSRF_INVALID",
              "message": "请求安全令牌已失效",
              "requestId": "01J3EXAMPLE00000000000001",
              "retryable": true,
              "details": {}
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "The caller exceeded an entry-specific rate limit",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
...
      "InternalProblem": {
        "description": "Unexpected internal failure",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "UuidV4": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "example": "4b9e8c22-7d38-4e4b-a1b0-7a0e1656e001"
      },
      "PublicUid": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UuidV4"
          }
        ],
        "description": "Stable public identity; never an internal BIGINT primary key"
      },
...
      "UtcTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\\.[0-9]{3}Z$",
        "example": "2026-07-26T03:15:20.123Z"
      },
...
      "ExpectedVersion": {
        "type": "integer",
        "minimum": 0,
        "description": "Last server version explicitly observed by the operator",
        "example": 3
      },
...
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId",
          "retryable",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*(\\.[A-Z][A-Z0-9_]*)+$",
            "example": "COMMON.IDEMPOTENCY_KEY_CONFLICT"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "example": "相同操作标识已绑定到不同请求"
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "example": "01J3EXAMPLE00000000000001"
          },
          "retryable": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "example": {}
          }
        }
      },
...
      "CsrfToken": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "token",
          "headerName"
        ],
        "properties": {
          "token": {
            "type": "string",
            "minLength": 32,
            "maxLength": 512,
            "description": "Opaque token kept in JavaScript memory only",
            "example": "csrf-example-value-not-a-production-secret-0001"
          },
          "headerName": {
            "type": "string",
            "const": "X-CSRF-TOKEN"
          }
        }
      },
      "CsrfTokenEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "const": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/CsrfToken"
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
      "WebLoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "loginName",
          "password"
        ],
        "properties": {
          "loginName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "example": "pilot.operator"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "writeOnly": true,
            "example": "not-a-real-password"
          }
        }
      },
      "OrganizationSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationCode",
          "displayName"
        ],
        "properties": {
          "organizationCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "example": "HZ-PILOT"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "湖州试点机构"
          }
        }
      },
      "WebSession": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sessionUid",
          "accountType",
          "subjectUid",
          "displayName",
          "capabilities",
          "organizations",
          "expiresAt",
          "version",
          "authVersion"
        ],
        "properties": {
          "sessionUid": {
            "$ref": "#/components/schemas/UuidV4"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "PLATFORM_ADMIN",
              "TENANT_PRINCIPAL",
              "STAFF"
            ]
          },
          "subjectUid": {
            "$ref": "#/components/schemas/PublicUid"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "contactPhone": {
            "oneOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ]
          },
          "tenantCode": {
            "oneOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "capabilities": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9-]*\\.[a-z][a-z0-9-]*$"
            }
          },
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationSummary"
            }
          },
          "expiresAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "version": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "authVersion": {
            "$ref": "#/components/schemas/ExpectedVersion"
          }
        }
      },
      "WebSessionEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "const": "OK"
          },
          "data": {
            "$ref": "#/components/schemas/WebSession"
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
...
    }
  },
...
}